refactor: 提取翻译和文本处理方法到独立文件
- FmMain.Translation.cs: 翻译逻辑 (30 methods)
- FmMain.TextProcessing.cs: 文本排版/标点/格式化 (34 methods)
