+++ TrOCR.csproj
@@ -334,6 +334,12 @@
     <Compile Include="FmMain.OCR.MenuHandlers.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Include="FmMain.Translation.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="FmMain.TextProcessing.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="FmMain.cs">
       <SubType>Form</SubType>
     </Compile>
